pyproject.toml
[tool.poetry]
name = "square"
version = "0.1.0"
description = "A wrapper for the Square API"
authors = ["Genetical <G3netical@protonmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.25.1"
pycountry = "^20.7.3"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"